I totally changed the way Display works.
Also, I am not putting any more 'Attempt' or 'contact' keywords in my Main Game Loop. This just makes everything messy.

Instead, it is possible to Display the trajectory with very short code in a for loop. (example is given in Attempt).
Also, I changed the step function. I am not sure, if the simulation will still work like this, but ... for the experimetnal trajectories I found a good solution.